添加登录注册

+++ app/models/account.py
@@ -8,7 +8,7 @@ class Account(BaseModel):
     password = db.Column(db.String(1024), nullable=False)
     country_code = db.Column(db.String(8), doc='国家代码', nullable=True)
     phone_number = db.Column(db.String(16), doc='手机号', nullable=True, index=True)
-    last_login_at = db.Column(db.Integer, nullable=True, doc='上次登录时间')
+    last_login_at = db.Column(db.BigInteger, nullable=True, doc='上次登录时间')
 
 
 class Profile(BaseModel):
